Update git & vim merge conflict settings.
Tell git not to create .orig backup files when performing merge conflict
resolution, so that I don't have to delete them every time I'm done with
a merge conflict!

Use diff3 to format merge conflict markers, so that the merged common
ancestor is shown with |||||||. It's important to be able to see the
merged common ancestor to have a complete picture of how to fix a merge
conflict.

Simplify vimdiff color scheme so that I can understand more clearly what
is going on when resolving merge conflicts with vimdiff and 3
side-by-side windows full of colors. This uses blue to highlight
conflicted sections and red to highlight the actual conflicted
characters.