Move personal info out of gitconfig.
Create gitconfig_local dotfile symlinked to ~/.gitconfig_local and then
add a [path] entry to the original gitconfig specifying
~/.gitconfig_local.

Place all personal information, user name, email, passwords, etc. into
~/.gitconfig_local instead of ~/.gitconfig.

Create .gitignore in dotfiles root to ignore gitconfig_local so that it
will never be accidentally committed to the dotfiles repository.

This is separate from ~/.extra, which is for bash, not git.