Add Firefox Pentadactyl settings.
Pentadactyl is a Firefox plugin that allows Vim-like keybindings in the
web browser. It's settings are defined via ~/.pentadactylrc.

Pentadactyl hides the Firefox toolbar by default in order to force you
to learn the keyboard shortcuts, which is generally a good idea.
However, after using Pentadactyl for a while, I have realized that the
Firefox toolbar is indispensable for a number of reasons:

- I like seeing the URL at the top of the page rather than the bottom.
- I like seeing the download progress bar in the downloads button.
- I have certain addons installed whose functionality would be very
  painful to access via Pentadactyl keyboard shortcuts
- Sometimes I just feel like being lazy and using the mouse ;-)

Create pentadactylrc and add a setting that shows the Firefox toolbar on
startup.