Move defaults write command to macos.sh.
970f688771a0adf00b33aa612cc565393d159ae9 "Fix disabling crash reporter
in .bash_profile." stopped attempting to disable the crash reporter
itself (not allowed with SIP enabled), which had to be disabled every
login in bash_profile, and instead just disabled the crash report
dialog, which is a defaults write command. This should now be placed in
macos.sh since it only needs to be run once.