Fix misleading vimrc comments.
Specify that plugins are only configured if the runtime directory of the
plugin exists; the current comments on conditional plugin loading say
"...if...was loaded by pathogen)" which implies the conditional check is
using pathogen to determine if a plugin has been loaded. However,
pathogen hasn't loaded any plugins yet during the sourcing of vimrc, so
this is incorrect and misleading.